Added Swagger and UTs

+++ account-service/src/main/java/com/neueda/test/account/VO/AccountDebitInfo.java
@@ -21,6 +21,6 @@ public class AccountDebitInfo {
 	@ToString.Exclude
 	private int pin;
 	
-	private int debitAmount;
+	private double debitAmount;
 
 }
